Update introduction.md
Remove duplicate words and fix typos on 

3. Single Request Format

In Moonlight architecture, all requests are made in a similar format. This makes it easier to understand and debug the requests. Requests can either be ~be~ of type JSON or form-data.

Every request must define the SERVICE and ACTION to ~exacute~ execute in the request body plus the ~rest~ rest of the payload as required by the service.

Advantages 6:
Frontend Integration :- One endpoint for all requests, one request format, one response format ~make~ makes it easy for frontend devs to integrate the api.